Added class "Poll" for the ligt posts,used PVectors similarly to the "Obstacle" class.
Adjusted size of canvas to be bigger.

Changed the position in where the barrels reset on the x-axis.

Due to the zise change, the class "Line" was changed so that the "for loop" was modified so that the lines extend aross the road.

Made it so that when the player presses "e" the car lights change to be random.

Modified obstacle speed and added a second barrel so that the gmae can be more challanging.